set triger product histories

+++ app/Http/Controllers/Api/ProductHistoryController.php
@@ -26,7 +26,22 @@ class ProductHistoryController extends Controller
                 ->leftjoin('employees as update', 'product_histories.updated_by', '=', 'update.id')
                 ->leftjoin('employees as delete', 'product_histories.deleted_by', '=', 'delete.id')
                 ->withTrashed()
-                ->get(['create.name as created_by_name', 'update.name as updated_by_name', 'delete.name as deleted_by_name', 'product_histories.*']);
+                ->get([
+                    'create.name as created_by_name',
+                    'update.name as updated_by_name',
+                    'delete.name as deleted_by_name',
+                    'products.unit',
+                    'product_histories.id',
+                    'product_histories.history_category',
+                    'product_histories.history_date',
+                    'product_histories.amount_of_product',
+                    'product_histories.product_price',
+                    'product_histories.product_expired_date',
+                    'product_histories.product_id',
+                    'product_histories.created_by',
+                    'product_histories.updated_by',
+                    'product_histories.deleted_by'
+                ]);
 
             if (count($productHistory) > 0) {
                 $response = [
@@ -49,51 +64,6 @@ class ProductHistoryController extends Controller
             $response = [
                 'status' => 'fails',
                 'message' => 'Mengambil Data Riwayat Produk Gagal -> Server Error',
-                'data' => null,
-            ];
-
-            return response()->json($response, Response::HTTP_INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        try {
-            $productHistory = ProductHistory::join('products', 'product_histories.product_id', '=', 'products.id')
-                ->leftjoin('employees as create', 'product_histories.created_by', '=', 'create.id')
-                ->leftjoin('employees as update', 'product_histories.updated_by', '=', 'update.id')
-                ->leftjoin('employees as delete', 'product_histories.deleted_by', '=', 'delete.id')
-                ->withTrashed()
-                ->where('product_histories.id', '=', $id)
-                ->get(['create.name as created_by_name', 'update.name as updated_by_name', 'delete.name as deleted_by_name', 'product_histories.*']);
-
-            if (count($productHistory) != 1) {
-                $response = [
-                    'status' => 'success',
-                    'message' => 'Mencari Data Riwayat Produk Sukses',
-                    'data' => $productHistory,
-                ];
-
-                return response()->json($response, Response::HTTP_OK);
-            }
-
-            $response = [
-                'status' => 'fails',
-                'message' => 'Mencari Data Riwayat Produk Gagal -> Data Kosong',
-                'data' => null,
-            ];
-
-            return response()->json($response, Response::HTTP_NOT_FOUND);
-        } catch (QueryException $e) {
-            $response = [
-                'status' => 'fails',
-                'message' => 'Mencari Data Riwayat Produk Gagal -> Server Error',
                 'data' => null,
             ];
 
@@ -125,7 +95,7 @@ class ProductHistoryController extends Controller
             'history_date' => $request->input('history_date'),
             'amount_of_product' => $request->input('amount_of_product'),
             'product_price' => $request->input('product_price'),
-            'total_price' => $request->input('total_price'),
+            'total_price' => $request->input('product_price') * $request->input('amount_of_product'),
             'product_expired_date' => $request->input('product_expired_date'),
             'product_id' => $request->input('product_id'),
             'created_by' => $request->input('created_by')
@@ -144,7 +114,7 @@ class ProductHistoryController extends Controller
         if ($validator->fails()) {
             $response = [
                 'status' => 'fails',
-                'message' => 'Menambah Data Riwayat Produk Gagal -> ' . $validator->errors(),
+                'message' => 'Menambah Data Riwayat Produk Gagal -> ' . $validator->errors()->first(),
                 'data' => null,
             ];
 
@@ -164,7 +134,7 @@ class ProductHistoryController extends Controller
         } catch (QueryException $e) {
             $response = [
                 'status' => 'fails',
-                'message' => 'Menambah Data Riwayat Produk Gagal -> Server Error',
+                'message' => 'Menambah Data Riwayat Produk Gagal -> Server Error ' . $e->getMessage(),
                 'data' => null,
             ];
 
@@ -181,11 +151,9 @@ class ProductHistoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $productHistory = ProductHistory::withTrashed()
-            ->where('id', '=', $id)
-            ->get();
+        $productHistory = ProductHistory::withTrashed()->find($id);
 
-        if (count($productHistory) != 1) {
+        if (is_null($productHistory)) {
             $response = [
                 'status' => 'fails',
                 'message' => 'Mengubah Data Riwayat Produk Gagal -> Data Tidak Ditemukan',
@@ -211,7 +179,7 @@ class ProductHistoryController extends Controller
             'history_date' => $request->input('history_date'),
             'amount_of_product' => $request->input('amount_of_product'),
             'product_price' => $request->input('product_price'),
-            'total_price' => $request->input('total_price'),
+            'total_price' => $request->input('product_price') * $request->input('amount_of_product'),
             'product_expired_date' => $request->input('product_expired_date'),
             'product_id' => $request->input('product_id'),
             'updated_by' => $request->input('updated_by')
@@ -230,7 +198,7 @@ class ProductHistoryController extends Controller
         if ($validator->fails()) {
             $response = [
                 'status' => 'fails',
-                'message' => 'Mengubah Data Riwayat Produk Gagal -> ' . $validator->errors(),
+                'message' => 'Mengubah Data Riwayat Produk Gagal -> ' . $validator->errors()->first(),
                 'data' => null,
             ];
 
@@ -261,16 +229,15 @@ class ProductHistoryController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request, $id)
     {
-        $productHistory = ProductHistory::withTrashed()
-            ->where('id', '=', $id)
-            ->get();
+        $productHistory = ProductHistory::withTrashed()->find($id);
 
-        if (count($productHistory) != 1) {
+        if (is_null($productHistory)) {
             $response = [
                 'status' => 'fails',
                 'message' => 'Menghapus Data Riwayat Produk Gagal -> Data Tidak Ditemukan',
@@ -281,12 +248,19 @@ class ProductHistoryController extends Controller
         }
 
         try {
-            $input = [
-                'deleted_by' => $request->input('deleted_by')
-            ];
-
-            $productHistory->update($input);
-            $productHistory->delete();
+            if (is_null($productHistory->deleted_at)) {
+                $input = [
+                    'deleted_by' => $request->input('deleted_by')
+                ];
+                $productHistory->update($input);
+                $productHistory->delete();
+            } else {
+                $input = [
+                    'deleted_by' => null
+                ];
+                $productHistory->update($input);
+                $productHistory->restore();
+            }
 
             $response = [
                 'status' => 'success',
